found a solution for writing out framedata
